#1 🔨 refactor: Layout , SideMenu FSD 폴더 구조 및 CSS 분리
