+++ src/app/index.ts
@@ -1,0 +1,2 @@
+export { default as App } from './App';
+export { default } from './main';
