+++ src/app/main.tsx
@@ -2,8 +2,8 @@ import '@ant-design/v5-patch-for-react-19';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import './index.css';
-import App from '@/App';
+import './styles';
+import App from './App';
 import { configureMessage } from '@/shared/config/messageConfig';
 
 // 토스트 메시지 전역 설정
