+++ src/app/styles/index.ts
@@ -1,0 +1,2 @@
+import './index.css';
+import './App.css';
