+++ src/app/vite-env.d.ts
