+++ src/entities/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './store';
+export * from './model';
