+++ src/entities/auth/model/auth.ts
