+++ src/entities/auth/store/index.ts
@@ -1,0 +1,1 @@
+export { useAuthStore } from './authStore';
