+++ src/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './admin';
+export * from './auth';
