+++ src/features/auth/ui/LoginForm/LoginForm.tsx
@@ -1,7 +1,7 @@
 import { Form, Input, Button, Card, Typography, message } from 'antd';
 import { UserOutlined, LockOutlined } from '@ant-design/icons';
 import { useAuthStore } from '@/entities/auth/store/authStore';
-import type { LoginFormData } from '@/shared/types/auth';
+import type { LoginFormData } from '@/entities/auth';
 import * as styles from './LoginForm.css';
 
 const { Title, Text } = Typography;
