+++ src/shared/index.ts
@@ -1,2 +1,5 @@
-export * from './types/auth';
-export * from './config/messageConfig';
+export * from './ui';
+export * from './config';
+export * from './scripts';
+export * from './model';
+export * from './store';
