+++ src/shared/scripts/index.ts
@@ -1,0 +1,1 @@
+export * from './commitScripts';
