+++ src/shared/store/index.ts
@@ -1,0 +1,1 @@
+export { useLayoutStore } from './layoutStore';
