+++ src/shared/ui/Layout/index.ts
@@ -1,0 +1,1 @@
+export { AppLayout } from './Layout';
