+++ src/shared/ui/SideMenu/index.ts
@@ -1,2 +1,1 @@
 export { SideMenu } from './SideMenu';
-export { AppLayout } from './Layout';
