+++ src/widgets/ActivityFeed/index.ts
@@ -1,0 +1,1 @@
+export { ActivityFeed } from './ActivityFeed';
