+++ src/widgets/DashboardStats/index.ts
@@ -1,0 +1,1 @@
+export { DashboardStats } from './DashboardStats';
