+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from './DashboardStats';
+export * from './ActivityFeed';
+export * from './UserProfile';
