+++ tsconfig.app.json
@@ -29,5 +29,5 @@
 			"@/*": ["src/*"]
 		}
 	},
-	"include": ["src"]
+	"include": ["src", "src/app/vite-env.d.ts"]
 }
