+++ vite.config.ts
@@ -6,6 +6,13 @@ import path from 'path';
 // https://vite.dev/config/
 export default defineConfig({
 	plugins: [react(), vanillaExtractPlugin()],
+	build: {
+		rollupOptions: {
+			input: {
+				main: path.resolve(__dirname, 'index.html'),
+			},
+		},
+	},
 	resolve: {
 		alias: {
 			'@': path.resolve(__dirname, './src'),
